RAG_Pipeline/requirements.txt
streamlit==1.31.1

requests==2.32.5
beautifulsoup4==4.12.3
lxml==6.0.2

ddgs==9.10.0

sentence-transformers==2.2.2
faiss-cpu==1.8.0


transformers==4.30.2
torch==2.0.1
tokenizers==0.13.3
sentencepiece==0.1.99
safetensors==0.7.0

numpy==1.26.4
scikit-learn==1.7.2
scipy==1.15.3

tqdm==4.67.2
huggingface-hub==0.16.4
